backend/requirements.txt
fastapi
uvicorn
gitpython
pydantic
google-auth
requests
